+++ vscode/src/chat/chat-view/ChatPanelProvider.ts
@@ -2,15 +2,19 @@ import * as vscode from 'vscode'
 
 import { CodyPrompt, CustomCommandType } from '@sourcegraph/cody-shared/src/chat/prompts'
 import { ChatMessage, UserLocalHistory } from '@sourcegraph/cody-shared/src/chat/transcript/messages'
-import { CodeBlockMeta } from '@sourcegraph/cody-ui/src/chat/CodeBlocks'
 
 import { View } from '../../../webviews/NavBar'
-import { getActiveEditor } from '../../editor/active-editor'
 import { logDebug } from '../../log'
 import { telemetryService } from '../../services/telemetry'
 import { telemetryRecorder } from '../../services/telemetry-v2'
 import { createCodyChatTreeItems } from '../../services/treeViewItems'
 import { TreeViewProvider } from '../../services/TreeViewProvider'
+import {
+    handleCodeFromInsertAtCursor,
+    handleCodeFromSaveToNewFile,
+    handleCopiedCode,
+} from '../../services/utils/codeblock-action-tracker'
+import { openExternalLinks, openFilePath, openLocalFileWithRange } from '../../services/utils/workspace-action'
 import { MessageErrorType, MessageProvider, MessageProviderOptions } from '../MessageProvider'
 import { ExtensionMessage, getChatModelsForWebview, WebviewMessage } from '../protocol'
 
@@ -72,42 +76,32 @@ export class ChatPanelProvider extends MessageProvider {
                 this.transcript.setChatModel(message.model)
                 break
             case 'executeRecipe':
-                await this.setWebviewView('chat')
                 await this.executeRecipe(message.recipe, '', 'chat')
                 break
+            case 'custom-prompt':
+                await this.onCustomPromptClicked(message.title, message.value)
+                break
             case 'insert':
-                await this.handleInsertAtCursor(message.text, message.metadata)
+                await handleCodeFromInsertAtCursor(message.text, message.metadata)
                 break
             case 'newFile':
-                await this.handleSaveToNewFile(message.text, message.metadata)
+                handleCodeFromSaveToNewFile(message.text, message.metadata)
+                await this.editor.createWorkspaceFile(message.text)
                 break
             case 'copy':
-                await this.handleCopiedCode(message.text, message.eventType, message.metadata)
+                await handleCopiedCode(message.text, message.eventType === 'Button', message.metadata)
                 break
             case 'event':
                 telemetryService.log(message.eventName, message.properties)
                 break
             case 'links':
-                void this.openExternalLinks(message.value)
-                break
-            case 'custom-prompt':
-                await this.onCustomPromptClicked(message.title, message.value)
+                void openExternalLinks(message.value)
                 break
             case 'openFile':
-                await this.openFilePath(message.filePath)
+                await openFilePath(message.filePath, this.webviewPanel?.viewColumn)
                 break
             case 'openLocalFileWithRange':
-                await this.openLocalFileWithRange(
-                    message.filePath,
-                    message.range
-                        ? new vscode.Range(
-                              message.range.startLine,
-                              message.range.startCharacter,
-                              message.range.endLine,
-                              message.range.endCharacter
-                          )
-                        : undefined
-                )
+                await openLocalFileWithRange(message.filePath, message.range)
                 break
             default:
                 this.handleError('Invalid request type from Webview Panel', 'system')
@@ -212,57 +206,6 @@ export class ChatPanelProvider extends MessageProvider {
         void this.webview?.postMessage({ type: 'errors', errors: errorMsg })
     }
 
-    /**
-     * Handles insert event to insert text from code block at cursor position
-     * Replace selection if there is one and then log insert event
-     * Note: Using workspaceEdit instead of 'editor.action.insertSnippet' as the later reformats the text incorrectly
-     */
-    private async handleInsertAtCursor(text: string, meta?: CodeBlockMeta): Promise<void> {
-        const selectionRange = getActiveEditor()?.selection
-        const editor = getActiveEditor()
-        if (!editor || !selectionRange) {
-            this.handleError('No editor or selection found to insert text', 'system')
-            return
-        }
-
-        const edit = new vscode.WorkspaceEdit()
-        // trimEnd() to remove new line added by Cody
-        edit.insert(editor.document.uri, selectionRange.start, text + '\n')
-        await vscode.workspace.applyEdit(edit)
-
-        // Log insert event
-        const op = 'insert'
-        const eventName = op + 'Button'
-        this.editor.controllers.inline?.setLastCopiedCode(text, eventName, meta?.source, meta?.requestID)
-    }
-
-    /**
-     * Handles insert event to insert text from code block to new file
-     */
-    private async handleSaveToNewFile(text: string, meta?: CodeBlockMeta): Promise<void> {
-        // Log insert event
-        const op = 'save'
-        const eventName = op + 'Button'
-        this.editor.controllers.inline?.setLastCopiedCode(text, eventName, meta?.source, meta?.requestID)
-
-        await this.editor.createWorkspaceFile(text)
-    }
-
-    /**
-     * Handles copying code and detecting a paste event.
-     * @param text - The text from code block when copy event is triggered
-     * @param eventType - Either 'Button' or 'Keydown'
-     */
-    private async handleCopiedCode(text: string, eventType: 'Button' | 'Keydown', meta?: CodeBlockMeta): Promise<void> {
-        // If it's a Button event, then the text is already passed in from the whole code block
-        const copiedCode = eventType === 'Button' ? text : await vscode.env.clipboard.readText()
-        const eventName = eventType === 'Button' ? 'copyButton' : 'keyDown:Copy'
-        // Send to Inline Controller for tracking
-        if (copiedCode) {
-            this.editor.controllers.inline?.setLastCopiedCode(copiedCode, eventName, meta?.source, meta?.requestID)
-        }
-    }
-
     protected handleCodyCommands(prompts: [string, CodyPrompt][]): void {
         void this.webview?.postMessage({
             type: 'custom-prompts',
@@ -311,52 +254,6 @@ export class ChatPanelProvider extends MessageProvider {
         }
         await this.clearHistory()
         this.webviewPanel?.dispose()
-    }
-
-    /**
-     * Open file in editor or in sourcegraph
-     */
-    protected async openFilePath(filePath: string): Promise<void> {
-        const rootUri = this.editor.getWorkspaceRootUri()
-        if (!rootUri) {
-            this.handleError('Failed to open file: missing rootUri', 'system')
-            return
-        }
-        try {
-            const doc = await vscode.workspace.openTextDocument(vscode.Uri.joinPath(rootUri, filePath))
-            let viewColumn = vscode.ViewColumn.Beside
-            // Open file next to current webview panel column
-            if (this.webviewPanel?.viewColumn) {
-                viewColumn = this.webviewPanel.viewColumn - 1 || this.webviewPanel.viewColumn + 1
-            }
-            await vscode.window.showTextDocument(doc, { viewColumn, preserveFocus: false })
-        } catch {
-            // Try to open the file in the sourcegraph view
-            const sourcegraphSearchURL = new URL(
-                `/search?q=context:global+file:${filePath}`,
-                this.contextProvider.config.serverEndpoint
-            ).href
-            void this.openExternalLinks(sourcegraphSearchURL)
-        }
-    }
-
-    /**
-     * Open file in editor (assumed filePath is absolute) and optionally reveal a specific range
-     */
-    protected async openLocalFileWithRange(filePath: string, range?: vscode.Range): Promise<void> {
-        const doc = await vscode.workspace.openTextDocument(vscode.Uri.file(filePath))
-        await vscode.window.showTextDocument(doc, { selection: range })
-    }
-
-    /**
-     * Open external links
-     */
-    private async openExternalLinks(uri: string): Promise<void> {
-        try {
-            await vscode.env.openExternal(vscode.Uri.parse(uri))
-        } catch (error) {
-            throw new Error(`Failed to open file: ${error}`)
-        }
     }
 
     /**